+++ handshake/crypto_setup_client_test.go
@@ -71,7 +71,7 @@ func (m *mockCertManager) Verify(hostname string) error {
 	return m.verifyError
 }
 
-var _ = Describe("Crypto setup", func() {
+var _ = Describe("Client Crypto Setup", func() {
 	var cs *cryptoSetupClient
 	var certManager *mockCertManager
 	var stream *mockStream
@@ -81,7 +81,7 @@ var _ = Describe("Crypto setup", func() {
 	BeforeEach(func() {
 		shloMap = map[Tag][]byte{
 			TagPUBS: []byte{0x0, 0x1, 0x2, 0x3, 0x4, 0x5, 0x6, 0x7, 0x8, 0x9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf, 0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d, 0x1e, 0x1f},
-			TagVER:  protocol.SupportedVersionsAsTags,
+			TagVER:  []byte{},
 		}
 		keyDerivation := func(forwardSecure bool, sharedSecret, nonces []byte, connID protocol.ConnectionID, chlo []byte, scfg []byte, cert []byte, divNonce []byte, pers protocol.Perspective) (crypto.AEAD, error) {
 			keyDerivationCalledWith = &keyDerivationValues{