+++ session_test.go
@@ -149,6 +149,7 @@ var _ = Describe("Session", func() {
 			0,
 			scfg,
 			func(Session, bool) {},
+			nil,
 		)
 		Expect(err).NotTo(HaveOccurred())
 		sess = pSess.(*session)
@@ -183,6 +184,7 @@ var _ = Describe("Session", func() {
 				0,
 				scfg,
 				func(Session, bool) {},
+				nil,
 			)
 			Expect(err).ToNot(HaveOccurred())
 			Expect(*(*[]byte)(unsafe.Pointer(reflect.ValueOf(sess.(*session).cryptoSetup).Elem().FieldByName("sourceAddr").UnsafeAddr()))).To(Equal([]byte{192, 168, 100, 200}))
@@ -198,6 +200,7 @@ var _ = Describe("Session", func() {
 				0,
 				scfg,
 				func(Session, bool) {},
+				nil,
 			)
 			Expect(err).ToNot(HaveOccurred())
 			Expect(*(*[]byte)(unsafe.Pointer(reflect.ValueOf(sess.(*session).cryptoSetup).Elem().FieldByName("sourceAddr").UnsafeAddr()))).To(Equal([]byte("192.168.100.200:1337")))