+++ Changelog.md
@@ -2,4 +2,5 @@
 
 ## v0.6.0 (unreleased)
 
+- Add a `quic.Config` option for QUIC versions
 - Various bugfixes